修复oracle一张表最多只能有一个long类型的字段，e_upms_operate_log创建会失败问题

+++ erupt-security/src/main/java/xyz/erupt/security/interceptor/EruptSecurityInterceptor.java
@@ -186,7 +186,7 @@ public class EruptSecurityInterceptor extends WebRequestHandlerInterceptorAdapte
                         operate.setReqParam(findRequestParamVal(request));
                     }
                     operate.setCreateTime(new Date());
-                    operate.setTotalTime(operate.getCreateTime().getTime() - RequestBodyTL.get().getDate());
+                    operate.setTotalTime((int) (operate.getCreateTime().getTime() - RequestBodyTL.get().getDate()));
                     RequestBodyTL.remove();
                     if (null != ex) {
                         operate.setErrorInfo(ExceptionUtils.getStackTrace(ex));
